fix(file size): use Long instead of float
Float only stores up to 16 million, and file size is stored in bytes. A video can easily be over 16MB (16 million bytes), therefore Long is needed.

+++ src/main/java/com/ddf/vodsystem/entities/Vod.java
@@ -42,7 +42,7 @@ public class Vod {
     private Float duration;
 
     @Column(name = "file_size", nullable = false)
-    private Float fileSize;
+    private Long fileSize;
 
     @Column(name = "video_path", nullable = false, length = 255)
     private String videoPath;
